move is_cg to core.function_spaces
